webadmin/ebscab/requirements.txt
amqplib==1.0.2
anyjson==0.3.3
BeautifulSoup==3.2.1
billiard==2.7.3.33
celery-with-redis==3.0
django-autocomplete-light==2.3.3
django-crispy-forms==1.6.1
django-filter==1.1.0
django-mathfilters==0.4.0
django-selectable==1.0.0
django-simple-captcha==0.5.5
django-tagging==0.4.5
django==1.11.7
ecdsa==0.9
ipaddr==2.1.10
IPy==0.81
jsonfield==2.0.2
kombu==3.0.37
lxml==3.8.0
m3-cdecimal==2.3
Mako==0.9.0
MarkupSafe==0.18
openpyxl==1.6.2
paramiko==1.12.0
psycopg2==2.5.4
pycrypto==2.6
pyExcelerator==0.6.4.1
pylibmc==1.2.3
pyrad==2.0
python-cjson==1.0.5
python-dateutil==2.1
python-memcached
pytils==0.2.3
pytz==2017.2
setuptools==36.2.7
simplejson==3.3.1
six==1.4.1
SQLAlchemy==0.8.2
Twisted==13.1.0
tzlocal==1.4
wsgiref==0.1.2
zope.interface==4.0.5
jinja2

